Linter file renaming to better reflect its current intent
